+++ hasklepias-main/ChangeLog.md
@@ -1,5 +1,9 @@
 # Changelog for hasklepias-main
 
+## 0.22.2
+
+* Adds tests for the example application.
+
 ## 0.22.1
 
 * Updates examples to work with `hasklepias-core` `0.22.0`.